Compose Cancel test not compile/run on windows (no signals)
